chore: add idea of '--msdos-volume' and '--msdos-device' as possible substitutes for '-f' on MS-DOS and Windows, to select raw device and volume
On MS-DOS, it is impossible to select a raw volume or raw device (ways
to treat devices as file-like), so the option "--msdos-volume" would
select a volume, maybe by syntax "A:", and "ghdisk.fat" would I/O the
the volume. The option "--msdos-device" would be similar, but instead of
the volume, it would select the device. For diskettes, both would work.

This can be useful to debug and explore MS-DOS FAT volumes live.