build: add MSVC support for building
The Makefiles uses defines to define a compiler-agnostic way to compile
C files and object files. A notable difference is that GCC/Clang compile
to binary without creating an object file, but MSVC (apparently) cannot
do that. Hence, it is always necessary to first compile to an object
file.

Further information is given on 'README.MD'.