feat: basic ad-hoc command-line parsing with no command
The current implementation is meant to be temporary, and grow into its
own file.

If nothing is given, the usage message is printed with success.
Otherwise, the syntax is "<options> <command> [...]", which restricts
that commands cannot start with "-".

The options are help options ("--help" or "-h") or version options
("--version" or "-v"). None of them take an argument. If only the
version option is set, then the version message is printed with success;
if only the help option is set, then the usage message is printed with
success. If both are set, the help message takes preference.

Short option syntax is implemented, such that each letter is one option,
and the next is either its argument or the next option. For example, the
argument "-vah" is equivalent to "-v -a -h".

The equal sign for separating option name and argument is not implemented, yet..