feat: multilingual support with static, embedded strings and replacement
Translatable strings are written by the programmer as '_("Hello, World!\n")',
where '_' is function which just returns its input. If the project is built
without translation, '_()' acts as above.

If the project is built with translation using the Make user parameter
'M_LANG', the C file 'a.c' is preprocessed to 'a.i', the text matches of
'_(STR)' are replaced with the translated string in 'po/<M_LANG>/a.po', by the
Python script 'lang.py'. For example, the text 'printf(_("Hello, World!\n"))'
is replaced with 'printf("Ola, Mundo!\n")'. Hence, the choice of language is
singular and not dynamic, but static.

The system is somewhat based on GNU 'xgettext', but uses the Python script
'lang.py <command>', which needs Python >=3.0. The script create a template
from one source code file, using regex matches and the command 'create_pot'.
The language-specific file is created by simply copying the template file. The
messages are translated manually. To replace the translatable strings in the
preprocessed files, the command 'replace_c_file'. To update the template and
translation files, by adding new translatable messages and removing useless
ones, the commands 'update_po' and 'update_pot' are used. To check if the POT
and PO files are updated, the option '-n' is used with the previous two
commands; if not up-to-date, the script exist with a non-zero value, otherwise
it exits with 0.

The template file (called POT file) and the language-specific translation file
(called PO file) use the exact same syntax, which is _not the same_ as that of
'xgettext'; however, it is based upon that. For reference, simply check
'po/pt/ghdisk.fat.po'. There is no header, because... Whatever. It uses ASCII,
and technically supports UTF-8, but it should only use ASCII.

The Makefile asserts the POT and PO files are up-to-date.

The script 'lang.py' is open to change at any time, for fixes and features,
since its use is specific to this project, and is made ad-hoc.