build: refactor 'lang.py' for PO/POT file parser to be line-based, instead of pure regex
Before, the function 'parse_pot()' parsed by first finding the matches
of the pair of 'msgid' and 'msgstr' through regex, then it parsed and
verified validity. This is problematic, because regex is kinda slow and
when a PO file was incorrect, it frequently took too long. Also, this
adds more flexibility and should make the code more readable.