build: fix fatal error in CL for multilingual support, where preprocessed files ending in '.i'
The compiler 'CL' compiler uses the file extension to detect the file
type, by default. The file extension ".i" is unknown to MSVC and when a
file type is unknown by extension, it is assumed by 'CL' to be an object
file.

The file extension ".i" was replaced with ".c".