refactor: initiate console in real entry function to enable printing error and add error messages
The function 'sys_init()' only initialised the standard streams
(specifically, only on Windows NT), and so not another function was
created. Instead, 'sys_init()' was renamed to 'sys_init_console()', its
use removed from 'ghdisk_fat()', and used in 'main/*'. It does not
return fail because, as usual, programs are not taken responsible to
make sure that printing worked. However, this may be changed later...

In 'main/nt.c', a few errors messages were added, with translation on
'po/'. It presented that the current directory structure of 'po/' is
problematic.

No system-specific error message is printed. This is to be supported
shortly.